requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
requests==2.31.0
web3==6.15.1
